refactor(wallets): move on-chain setup service into plugin
