+++ BTCPayServer/Hosting/BTCPayServerServices.cs
@@ -361,7 +361,6 @@ namespace BTCPayServer.Hosting
             services.TryAddSingleton<StoreLabelRepository>();
             services.TryAddSingleton<PaymentRequestRepository>();
             services.TryAddSingleton<BTCPayWalletProvider>();
-            services.TryAddSingleton<OnChainWalletSetupService>();
             services.AddSingleton<PendingTransactionService>();
             services.AddScheduledTask<PendingTransactionService>(TimeSpan.FromMinutes(10));
             // PendingTransactionWebhookProvider webhooks registered in WebhookExtensions
