+++ BTCPayServer/Plugins/Multisig/Controllers/UIMultisigSetupController.cs
@@ -10,8 +10,8 @@ using BTCPayServer.Data;
 using BTCPayServer.Plugins.Multisig.Models;
 using BTCPayServer.Plugins.Multisig.Services;
 using BTCPayServer.Plugins.Wallets;
+using BTCPayServer.Plugins.Wallets.Services;
 using BTCPayServer.Services.Stores;
-using BTCPayServer.Services.Wallets;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Localization;
